requirements.txt
uvicorn[standard]==0.22.0
fastapi==0.95.2
sqlalchemy==2.0.23
sqlmodel==0.0.12
pandas==2.1.3
cryptography==41.0.7
python-multipart==0.0.6
psycopg2==2.9.9
pytest==7.4.3
pyjwt==2.8.0
requests==2.31.0